refactor: TossPaymentProperties 패키지 변경
